requirements.txt
oracledb
kafka-python